# apps/docs/public/openapi.yaml
openapi: '3.0.0'
info:
  version: v2.7.0/366f528
  title: Labrinth
  termsOfService: https://modrinth.com/legal/terms
  contact:
    name: Modrinth Support
    url: https://support.modrinth.com
    email: support@modrinth.com
  description: |
    This documentation doesn't provide a way to test our API. In order to facilitate testing, we recommend the following tools:

    - [cURL](https://curl.se/) (recommended, command-line)
    - [ReqBIN](https://reqbin.com/) (recommended, online)
    - [Postman](https://www.postman.com/downloads/)
    - [Insomnia](https://insomnia.rest/)
    - Your web browser, if you don't need to send headers or a request body

    Once you have a working client, you can test that it works by making a `GET` request to `https://staging-api.modrinth.com/`:

    ```json
    {
      "about": "Welcome traveler!",
      "documentation": "https://docs.modrinth.com",
      "name": "modrinth-labrinth",
      "version": "2.7.0"
    }
    ```

    If you got a response similar to the one above, you can use the Modrinth API!
    When you want to go live using the production API, use `api.modrinth.com` instead of `staging-api.modrinth.com`.

    ## Authentication
    This API has two options for authentication: personal access tokens and [OAuth2](https://en.wikipedia.org/wiki/OAuth).
    All tokens are tied to a Modrinth user and use the `Authorization` header of the request.

    Example:
    ```
...
    ```

...
    - those which create data (such as version creation)
...

...

...

...
servers:
...
components:
  parameters:
...
    MultipleProjectIdentifier:
      in: query
      name: ids
      description: The IDs and/or slugs of the projects
      schema:
        type: string
        example: '["AABBCCDD", "EEFFGGHH"]'
      required: true
...
  schemas:
...
    BaseProject:
      type: object
      properties:
        slug:
          type: string
          description: "The slug of a project, used for vanity URLs. Regex: ```^[\\w!@$()`.+,\"\\-']{3,64}$```"
          example: my_project
        title:
          type: string
          description: The title or name of the project
          example: My Project
        description:
          type: string
          description: A short description of the project
          example: A short description
        categories:
          type: array
          items:
            type: string
          example: [technology, adventure, fabric]
          description: A list of the categories that the project has
        client_side:
          type: string
          enum: [required, optional, unsupported, unknown]
          description: The client side support of the project
          example: required
        server_side:
          type: string
          enum: [required, optional, unsupported, unknown]
          description: The server side support of the project
          example: optional
    # Fields added to search results and direct project lookups that cannot be edited.
    ServerRenderedProject:
      allOf:
        - $ref: '#/components/schemas/BaseProject'
        - type: object
          properties:
            project_type:
              type: string
              enum: [mod, modpack, resourcepack, shader]
              description: The project type of the project
              example: mod
            downloads:
              type: integer
              description: The total number of downloads of the project
            icon_url:
              type: string
              example: https://cdn.modrinth.com/data/AABBCCDD/b46513nd83hb4792a9a0e1fn28fgi6090c1842639.png
              description: The URL of the project's icon
              nullable: true
            color:
              type: integer
              example: 8703084
              description: The RGB color of the project, automatically generated from the project icon
              nullable: true
            thread_id:
              type: string
              example: TTUUVVWW
              description: The ID of the moderation thread associated with this project
            monetization_status:
              type: string
              enum: [monetized, demonetized, force-demonetized]
          required:
            - project_type
            - downloads
    # The actual result in search.
...
    NonSearchProject:
      allOf:
        - $ref: '#/components/schemas/BaseProject'
        - type: object
          properties:
            body:
              type: string
              description: A long form description of the project
              example: A long body describing my project in detail
            status:
              type: string
              enum:
                [
                  approved,
                  archived,
                  rejected,
                  draft,
                  unlisted,
                  processing,
                  withheld,
                  scheduled,
                  private,
                  unknown,
                ]
              description: The status of the project
              example: approved
            requested_status:
              type: string
              enum: [approved, archived, unlisted, private, draft]
              description: The requested status when submitting for review or scheduling the project for release
              nullable: true
            additional_categories:
              type: array
              items:
                type: string
              description: A list of categories which are searchable but non-primary
              example: [technology, adventure, fabric]
            issues_url:
              type: string
              description: An optional link to where to submit bugs or issues with the project
              example: https://github.com/my_user/my_project/issues
              nullable: true
            source_url:
              type: string
              description: An optional link to the source code of the project
              example: https://github.com/my_user/my_project
              nullable: true
            wiki_url:
              type: string
              description: An optional link to the project's wiki page or other relevant information
              example: https://github.com/my_user/my_project/wiki
              nullable: true
            discord_url:
              type: string
              description: An optional invite link to the project's discord
              example: https://discord.gg/AaBbCcDd
              nullable: true
            donation_urls:
              type: array
              items:
                $ref: '#/components/schemas/ProjectDonationURL'
              description: A list of donation links for the project
            wiki_open:
              type: boolean
              description: Whether the project has an open wiki
              default: false
            issues_type:
              type: integer
              description: The type of issues system (0=disabled, 1=internal, 2=external)
              default: 0
              enum: [0, 1, 2]
            forum:
              type: string
              description: The ID of the project's forum/discussion thread
              nullable: true
    ProjectDonationURL:
      type: object
      properties:
        id:
          type: string
          description: The ID of the donation platform
          example: patreon
        platform:
          type: string
          description: The donation platform this link is to
          example: Patreon
        url:
          type: string
          description: The URL of the donation platform and user
          example: https://www.patreon.com/my_user
    # Fields available only when editing or creating a project
...
    Project:
      allOf:
        - $ref: '#/components/schemas/NonSearchProject'
        - $ref: '#/components/schemas/ServerRenderedProject'
        - type: object
          properties:
            id:
              type: string
              example: AABBCCDD
              description: The ID of the project, encoded as a base62 string
            team:
              type: string
              example: MMNNOOPP
              description: The ID of the team that has ownership of this project
            body_url:
              type: string
              deprecated: true
              default: null
              description: The link to the long description of the project. Always null, only kept for legacy compatibility.
              example: null
              nullable: true
            moderator_message:
              $ref: '#/components/schemas/ModeratorMessage'
            published:
              type: string
              format: ISO-8601
              description: The date the project was published
            updated:
              type: string
              format: ISO-8601
              description: The date the project was last updated
            approved:
              type: string
              format: ISO-8601
              description: The date the project's status was set to an approved status
              nullable: true
            queued:
              type: string
              format: ISO-8601
              description: The date the project's status was submitted to moderators for review
              nullable: true
            followers:
              type: integer
              description: The total number of users following the project
            license:
              $ref: '#/components/schemas/ProjectLicense'
            versions:
              type: array
              items:
                type: string
              example: [IIJJKKLL, QQRRSSTT]
              description: A list of the version IDs of the project (will never be empty unless `draft` status)
            game_versions:
              type: array
              items:
                type: string
              example: ['1.19', '1.19.1', '1.19.2', '1.19.3']
              description: A list of all of the game versions supported by the project
            loaders:
              type: array
              items:
                type: string
              example: ['forge', 'fabric', 'quilt']
              description: A list of all of the loaders supported by the project
            gallery:
              type: array
              items:
                $ref: '#/components/schemas/GalleryImage'
              description: A list of images that have been uploaded to the project's gallery
          required:
            - id
            - team
            - published
            - updated
            - followers
            - title
            - description
            - categories
            - client_side
            - server_side
            - slug
            - body
            - status
    ModeratorMessage:
      deprecated: true
      type: object
      properties:
        message:
          type: string
          description: The message that a moderator has left for the project
        body:
          type: string
          description: The longer body of the message that a moderator has left for the project
          nullable: true
      nullable: true
      example: null
      description: A message that a moderator sent regarding the project
    ProjectLicense:
      type: object
      properties:
        id:
          type: string
          description: The SPDX license ID of a project
          example: LGPL-3.0-or-later
        name:
          type: string
          description: The long name of a license
          example: GNU Lesser General Public License v3 or later
        url:
          type: string
          description: The URL to this license
          nullable: true
      description: The license of the project
    GalleryImage:
      type: object
      nullable: true
      properties:
        url:
          type: string
          description: The URL of the gallery image
          example: https://cdn.modrinth.com/data/AABBCCDD/images/009b7d8d6e8bf04968a29421117c59b3efe2351a.png
        featured:
          type: boolean
          description: Whether the image is featured in the gallery
          example: true
        title:
          type: string
          description: The title of the gallery image
          example: My awesome screenshot!
          nullable: true
        description:
          type: string
          description: The description of the gallery image
          example: This awesome screenshot shows all of the blocks in my mod!
          nullable: true
        created:
          type: string
          format: ISO-8601
          description: The date and time the gallery image was created
        ordering:
          type: integer
          description: The order of the gallery image. Gallery images are sorted by this field and then alphabetically by title.
          example: 0
      required:
        - url
        - featured
        - created
...
    PatchProjectsBody:
      type: object
      properties:
        categories:
          description: Set all of the categories to the categories specified here
          type: array
          items:
            type: string
        add_categories:
          description: Add all of the categories specified here
          type: array
          items:
            type: string
        remove_categories:
          description: Remove all of the categories specified here
          type: array
          items:
            type: string
        additional_categories:
          description: Set all of the additional categories to the categories specified here
          type: array
          items:
            type: string
        add_additional_categories:
          description: Add all of the additional categories specified here
          type: array
          items:
            type: string
        remove_additional_categories:
          description: Remove all of the additional categories specified here
          type: array
          items:
            type: string
        donation_urls:
          description: Set all of the donation links to the donation links specified here
          type: array
          items:
            $ref: '#/components/schemas/ProjectDonationURL'
        add_donation_urls:
          description: Add all of the donation links specified here
          type: array
          items:
            $ref: '#/components/schemas/ProjectDonationURL'
        remove_donation_urls:
          description: Remove all of the donation links specified here
          type: array
          items:
            $ref: '#/components/schemas/ProjectDonationURL'
        issues_url:
          type: string
          description: An optional link to where to submit bugs or issues with the projects
          example: https://github.com/my_user/my_project/issues
          nullable: true
        source_url:
          type: string
          description: An optional link to the source code of the projects
          example: https://github.com/my_user/my_project
          nullable: true
        wiki_url:
          type: string
          description: An optional link to the projects' wiki page or other relevant information
          example: https://github.com/my_user/my_project/wiki
          nullable: true
        discord_url:
          type: string
          description: An optional invite link to the projects' discord
          example: https://discord.gg/AaBbCcDd
          nullable: true
...
    InvalidInputError:
      type: object
      properties:
        error:
          type: string
          description: The name of the error
          example: 'invalid_input'
        description:
          type: string
          description: The contents of the error
          example: 'Error while parsing multipart payload'
      required:
        - error
        - description
    AuthError:
      type: object
      properties:
        error:
          type: string
          description: The name of the error
          example: 'unauthorized'
        description:
          type: string
          description: The contents of the error
          example: 'Authentication Error: Invalid Authentication Credentials'
      required:
        - error
        - description
    # Other
...
  securitySchemes:
    TokenAuth:
      type: apiKey
      in: header
      name: Authorization
...
paths:
...
  /projects:
    parameters:
      - $ref: '#/components/parameters/MultipleProjectIdentifier'
    get:
      summary: Get multiple projects
      operationId: getProjects
      tags:
        - projects
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    patch:
      summary: Bulk-edit multiple projects
      operationId: patchProjects
      tags:
        - projects
      security:
        - TokenAuth: ['PROJECT_WRITE']
      requestBody:
        description: Fields to edit on all projects specified
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchProjectsBody'
      responses:
        '204':
          description: Expected response to a valid request
        '400':
          description: Request was invalid, see given error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidInputError'
        '401':
          description: Incorrect token scopes or no authorization to access the requested item(s)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
...